requirements.txt
asgiref==3.9.1
cfgv==3.4.0
distlib==0.4.0
Django==5.2.5
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
identify==2.6.13
mypy==1.17.1
mypy_extensions==1.1.0
nodeenv==1.9.1
pathspec==0.12.1
platformdirs==4.4.0
pre_commit==4.3.0
PyJWT==2.10.1
PyYAML==6.0.2
ruff==0.12.9
sqlparse==0.5.3
types-PyYAML==6.0.12.20250809
typing_extensions==4.14.1
tzdata==2025.2
virtualenv==20.34.0
